Copy download to content URI
Add a new context menu item for downloaded files with the label "Share with other applications".

This functionality creates a new entry in Android's media database and copies the file to the public Downloads folder, where it may be viewed by other apps.

This is a suboptimal way to share downloaded content between Wolvic and other apps. It would be ideal to be able to store files in Android's content database by default and then open them in Wolvic, but the problem is that at the moment GeckoView is unable to handle Android's content URIs (which are all that the current Android APIs provide).

Fixes #207