[Extensions] Allow installation of locally downloaded extensions/addons (#391)
Fixes #22

This PR adds a Developer Setting to Allow installation of locally downloaded extensions.

Flow:
1. User enables Developer Option to Allow local addon installation (Settings > Developer Options > 
"Allow Installation of Local Addons")
2. User downloads .xpi file from the extension's website
3. User clicks on the corresponding entry in the Download Manager
4. The addon is installed and enabled

The installation flow for a LocalExtension is quite different from a regular addon. The primary difference
is that for local extensions, we do not have an "Addon" class instance available prior to installation when
we refresh our addons list, as info related to that is populated from the extension manifest only after 
installation. So the local addon installation needs to happen similar to how BuiltInExtensions are installed 
in the existing code.