OpenXR: Emulate Oculus Touch interaction profile if hand tracking is enabled (#385)
Since we are emulating controller functionality for hand tracking, we
need to always have an active interaction profile, even if Wolvic is
launched without any controllers active.

This patch emulates the activation of OculusTouch controllers at the
beginig of the OpenXR session, if hand tracking support is detected
at runtime. This includes settting the button count on emulated profiles
during UpdateInteractionProfile(). This is needed because the button
count may never be set if actual controllers have never been active
(e.g, when using only hand tracking).

Co-authored-by: Eduardo Lima <elima@filetea.me>
Co-authored-by: Sergio Villar Senin <svillar@igalia.com>