[Pico] Do not create layers for some widgets
Pico's runtime currently only allow 16 simultaneous layers. That's enough
for showing 1 window. However as we add windows to the scene the amount
of layers easily exceed that value. That's why, while the runtime does not
support more, we have to limit the amount of layers we create somehow.

The easiest way is to select a set of widgets to be rendered using
geometry instead of layers. As the render quality of layers is much
higher, we select, small, not very important widgets to be rendered
using geometry. For now it seems to be enough to do it only for the
TopBarWidget, TitleBarWidget and the WebXRInterstitialWidget.