Fixes a crash generating download thumnails
Due to some other bug, some downloads are stored using names without
any extension. That was making getMimeTypeFromUrl() return null in
those cases.

The problem is that we were incorrectly doing a null check in the
value returned by MimeTypeMap.getFileExtensionFromUrl(). As the docs
mention, that method returns the empty string in case of not finding
an extension. As the empty string is not null we were never early
returning UNKNOWN_FILE_TYPE, and instead we returned null because
MimeTypeMap.getSingleton().getMimeTypeFromExtension() does indeed
return null if the extension is not recognized.