Do not leak swapchains on resize
Swapchains cannot be resized. That's why when we resize widgets (like
for example window widgets when entering/leaving fullscreen) what we
really do is:
  1. Create a new swapchain with the new size
  2. Once the new swapchain is ready replace the previous one and
     destroy it

That behaviour is sane in theory, however it does leak swapchains each
time we leave fullscreen mode (i.e. when resize is triggered). This is
why. The current code creates a new swapchain each time VRLayer::Resize
is called. The layer notifies the SurfaceChangedDelegate (in this case
a Widget) about that. The Widget then calls
VRBrowser::DispatchCreateWidgetLayer(). This method talks to Java via
JNI and among other things, it passes a callback to the Java world.
That callback is executed later (by native code, Java just
enqueues the execution of a Runnable) during the first draw.

The problem arises when Resize() is called multiple times before
the first draw call. In that case, only the last callback is called,
meaning that, the other callbacks will be leaked (and thus the
created swapchain they carry). The leaked callback is created
in VRBrowser::DispatchCreateWidgetLayer() and is released in
BrowserWorld::runCallbackNative.

In order to fix thatm we created a new class FinalizerRunnable,
implementing Runnable, which takes two Runnables, one is the
main one executed by run() and the other one is executed in
finalize() when GC destroys the object. The latter is implemented
with a call to a new native method deleteCallbackNative()
which destroys the previously allocated callback (std::function).
Note that deleteCallbackNative() is only executed in case the
callback was not previously executed, because in that case,
it should have been already destroyed.

As we depend on GC, resources won't be immediately destroyed. This
could cause temporary peaks in swapchain allocations but they will
be eventually removed. This not only fixes a huge leak, but it's
also critical for systems with a small limit of available android
surface swapchains (like Pico).

Fixes #405