Download blob:http:// URIs
Blob URIs have the form blob:http://example.com/... and can be used to represent data that has been generated locally and does not exist in the remote server.

This PR makes it possible to download a Blob URI as if it was a regular file. Fortunately, Gecko provides us with an InputStream pointing at the blob's data, so all that we have to do is save this data to a file and update the Downloads database.

Fixes https://github.com/Igalia/wolvic/issues/304