[OpenXR][XRlayers] Limit the amount of layers sent to the compositor
Platforms like Meta or Pico allow a maximum of 16 layers to be
submitted for a given frame to the compositor. It's unlikely
that we'll reach that limit with our current setup but it does
not hurt to prevent us from doing so. Note that if a layer is
skipped for a given frame it'll be very likely sent in any of
the followup frames.

We do the same for the OculusVR backend so it makes sense to do
it also for OpenXR.