Prevent crashes when opening unsupported dialogs
Some Web dialogs are not supported by Wolvic yet. These include the color picker, date/time picker, "share", etc. This PR prevents a crash when a website tries to open one of these dialogs.

This is a quick fix to ensure the stability of the app, we still need to implement these dialogs in the future.

Fixes #408