[OpenXR][XRLayers] Use reorient transform to position cylinders
Whenever we enter/leave WebXR experiences (and immersive videos)
we recenter the UI yaw so the 2D window is properly placed in front
of the user no matter where it was before entering the experience.

We use something called a reorient transform to properly rotate the
whole scene graph. However for the case of cylinder layers they
were not aware of that transform because we are not currently
using the transform coming from the widget to position them.

With this PR we basically take that reorient transform into
account when creating the rotation matrix that is passed to
the cylinder layers.

Fixes #414