Add support to skipping frames
The OpenXR backend might ask the caller not to render a given frame
by setting XrFrameWaitInfo.shouldRender to false when calling
xrWaitFrame(). The current code was not even ignoring that, but it was
actually asserting in those cases.

That's why we have added a mShouldRender attribute to DeviceDelegate
which will be used by BrowserWorld to decide whether or not to
draw content. For the non-OpenXR backends that attribute is generally
true except when there is an error in StartFrame().