[Pico] Use PNG files with sRGB colorspace for environments
Pico's OpenXR runtime (v5.2) does not currently support compressed
texture formats. This means that it won't be able to render any
of the current environments as they are stored in .KTX textures.

As a workaround we've added the textures in PNG format in
SRGB color space which is indeed supported by the runtime. We
should revert this change once the runtime provides support
for compressed texture formats.