New extension to get around DelightVR UA filtering
Originally developed by Christopher Robert Van Wiemeersch <cvan@users.noreply.github.com>

This extension was uploaded as a PR to Firefox Reality
https://github.com/MozillaReality/FirefoxReality/pull/1213 some time
ago to circumvent DelightVR user agent filtering. It was later
discarded by the simpler but less robust approach of having a huge
list of UA overrides. It was discarded because there were performance
concerns, as the extension checks the names of every JS file a page
loads. I haven't benchmarked that but it does not look like an actual
issue to me, but we'll have to eventually reassess it.