OpenXR: Use distance between joints to detect pinch events
Pico4 hand-tracking support for aim state extension doesn't provide
pinching events information, so this patch moves the implementation
to be entirely based on joint locations as provided by
XR_EXT_hand_tracking.

The distance between the thumb-tip joint and that of the rest of the
fingers is calculated, then if it is lower than a threshold then a
pinch event is detected. This threshold has been empirically
calculated. It works well for both Quest2 and Pico4.