OpenXR: Fixes several issues with hand-tracking support to enable it on Pico4
Unfortunately we still need to use XR_FB_hand_tracking_aim to resolve
beam and pointer transform. In a future revision we will get that
entirely from hand joint locations to avoid depending on a vendor
extension. Good news is that Pico4 implements it although they don't
seem to advertis it.