[OpenXR] Add support for rendering different textures per eye
When dealing with compositor layers, OpenXR allows clients to specify
whether a layer will be displayed in both eyes, or just in one of
them by using the XrEyeVisibility enum. So far, we have been using
XR_EYE_VISIBILITY_BOTH which allows us to use the same layer for both
eyes.

However there are cases in which we do want to show different parts
of a given layer to each eye. For example with side-by-side video,
in order to get the stereo effect, we should display the left half of
the video only to the left eye and the right half to the right eye.

This change sets the ground for those use cases by adding a new
attribute to VRLayer specifying whether or not the layer will be
rendered to both eyes. This information will be used by the
OpenXR backend to only update the required number of layers and
also to fill in the value for XrEyeVisibility.