[OpenXR] Add Meta Quest Pro 3D controller models
We're uploading the 3D controller models obtained from
https://github.com/immersive-web/webxr-input-profiles/tree/main/packages/assets/profiles/meta-quest-touch-pro

Detecting the actual controller to be used was not trivial though.
The problem is that OpenXR incorrectly identifies the Meta Quest Pro
as "Oculus Quest2". Other properties, even the graphical ones are
identical between Quest2 and Pro. So the only way we've found to
identity the Pro is by a specific extension that is only supported
in the Pro. This is indeed fragile but we cannot do it better until
Meta fixes their OpenXR implementation.

That's why when we detect that extension we replace the system name
in the OpenXR system properties to "Meta Quest Pro" so it matches
the newly added input profile for the Pro.