Enable and disable add-ons
This PR fixes a bug that made it impossible to enable and disable add-ons in a consistent manner.

It also adds a couple of additional checks to prevent problems when an add-on has a NULL `id` field.

And it fixes a small UI problem in the dialog that we show when an add-on is removed.