Fix a crash when exiting the application
It might happen that by the time a VRLayerSurface is freed, the
Java context has been already released. In those cases trying to
delete the global reference kept by the surface causes a crash
due to a null pointer exception. In those cases we just simply have
to bail out.