Add amazon-es search plugin to fix a crash
Mozilla's android components v75 look for a amazon-es.xml search
plugin specification file when retrieving the list of search engines
available for a spanish locale. However that file is not distributed
with the aforementioned version that we're using causing an exception
to be thrown during the Wolvic initialization process. This makes
the whole app to crash even before showing the splash screen.

We're adding that file directly to our searchplugins directory so
it's merged in the final package, and thus, can be found by the
android components code. The file was retrieved from the android
components repository.