Use a safer Firefox-based User Agent
We recently changed our UA to be in the form of "Gecko/x.x Wolvic/x.x".
This gives us more visibility but it's actually causing issues, as
many services out there doing UA filtering do not recognize our UA
yet. An example is the very popular DelightXR player (a WebXR video
player) which does not work with our current UA but will work if we
mention Firefox.

So in order to remain very compatible and visible at the same time
we have decided to update the UA so it is made of the Gecko's
default UA + "Wolvic/version".