Append "Wolvic/1.2" to the default GeckoView User-Agent
Our User-Agent will be the one returned by default by GeckoView, plus "Wolvic" and its version.

This is a compromise between compatibility and visibility: sites that are aware of Wolvic will be able to detect it, whereas the rest will be able to identify us as a GeckoView browser.

The Session object is updated when the "User-Agent mode" setting changes. Note that this does not always means that the page will be automatically reloaded, basically to prevent reloading all tabs when the user changes this value in the Settings dialog.