Batch updateWidget native calls
Whenever Android asks us to update some widget, we queue a native task to perform the update in the native thread. However, sometimes we get several update widget requests in a short period of time, which involves expensive cycles of creation&destruction of swapchains for those widgets backed up by layers. This might cause peaks in the number of active
swapchain something that could be problematic (Pico for example).

Since the native tasks are not cancellable, we are enqueuing those updateWidget() calls before calling the native methods. To do this, we keep a hashtable with widget handles and the new widget placement. After 250ms we go through that hash table and call the native methods. Any updateWidget call that happens during the 250ms budget will replace the previous one, thus preventing unnecessary updates.

Fixes #448