[OpenXR] Adjust meta's controllers positions
Meta's controllers positions (in Quest2 and QuestPro) are not 100%
accurate. They show a slight shift in the three axis when compared to
the location of the controllers when rendered by the system. This
can be seen after clicking on the right menu button of the controller.

It does not look like a bug in our OpenXR code since other flavours
using layers, like Pico, do not show this missplacement. That's why
we're "manually" fine tuning the position of the controllers.