Message when the "Upload file" dialog is empty
Add a message that will be displayed when the list of files to upload is empty.

The message is "No files available for upload".

This situation can happen because the user has not downloaded anything yet,
or because the <input> element is using a filter that excludes all of the
existing files.