Remove WAKE_LOCK and FOREGROUND_SERVICE permissions
WAKE_LOCK and FOREGROUND_SERVICE are two permissions requested by GeckoView
which are not really needed in VR environments.

Note that by default GeckoView expects these permissions to be available,
even if they are not actually going to be needed.
PR https://github.com/Igalia/gecko-dev/pull/4 prevents any crashes when these
permissions are not available.