[OpenXR] Recompute reorient matrix whenever reference space changes
Many platforms allow recentering the view via a long press in the
system menu of a controller (like the Oculus button for example).
However we don't normally have access to those buttons via OpenXR
(the input/system path is generally not available for developers).
This means that we cannot detect those long presses by accessing
the controller data.

However there is another way to do that and is by listening to
the XrEventDataReferenceSpaceChangePending event which might
occur "due to the user recentering the space explicitly,
or the runtime otherwise switching to a different space
definition" literally from specs.

We were already listening to it and reseting the firstPose
attribute of the device delegate. This means that we could
use that information to know about reorientation events
and proceed to recompute the reorientMatrix as response.