[OpenXR][XRLayers] Apply widgets UI yaw before their transform
When adjusting the position of the cylinder layers we need to take
into account the orientation of the "front" direction that can be
reset by the user (for example when doing a long press in the menu
button). We were actually already doing that, the problem is that
we were applying that **after** positioning the cylinder layers.
We should do that **before** doing that instead, as rotations
are not commutative.