[OpenXR] Implement DeviceDelegateOpenXR::UpdateDisplayRefreshRate()
The OculusVR backend was able to dynamically set the refresh rate
of the device. This was lost with the migration to OpenXR. The main
consequence is that the refresh rate in Meta devices was reduced from
90Hz to 72Hz. This effectively reduced the FPS from ~85 to 72.

Fortunately there is a Facebook extension in OpenXR that allows us
to query and set the refresh rate of the device.