Improves hand-tracking pointer target behavior
Right now when hand-tracking is active we render the same beam
and pointer target as when controllers are active. This is not
optimal because 1) it doesn't give visual feedback when e.g a
trigger action happens (index+thumb pinch), and 2) the beams
are arguably distracting.

In both Pico4 and Quest2 devices, for example, a bigger circle
is shown as hand-tracking aim target, and it changes size based
on a pinch gesture between index+thumb. When a pinch is detected,
the target circle changes color too.

This patch gets the behavior closer to that of Pico and Quest by
rendering a circle whose radius varies based on the distance
between index and thumb; and color changes to 'Azure' blue (the
background color for selected items) when the pinch action is
triggered.

In a future revision we can make the circle transparent, to imitate
even more the behavior of Pico and Quest.