[OpenXR] Pause media playing on unfocus
Whenever the system dialog is shown on top of Wolvic (like after
pressing the system button of the right controller in Meta,
Pico...) the application should stop the playback of any media.

We can achieve that by listening to application state changes.
When the menu is shown the app transitions from FOCUSED to VISIBLE
state. The difference in OpenXR is that the former gets user
input while the later don't.

This behaviour fits very well with what's stated in the specs
which literally say "The application should continue running
its frame loop, rendering and submitting its composition layers,
although it may wish to pause its experience, as users cannot
interact with the application at this time"