Fix OpenXR crash on Pico v5.4.0 (#510)
Version 5.4.0 of Pico fixes a bug in the previous versions, but now
our workaround for that bug was causing the app to crash on startup.

Additionally, v5.4.0 changed the system name from "Pico: PICO HMD"
to "PICO 4", which broke the controller mappings.

Created a new header file SystemUtils.h which will contain different
utilities related to the current system. For now, SystemUtils.h contains 
the functions required to obtain, parse and compare the system's 
Build ID version.