Handle downloaded file URIs correctly
This PR implements additional functionality in the Download object to prevent errors when handling encoded and decoded file URIs:
- getOutputFileUri: returns a "file://" URI;
- getOutputFileUriAsString: returns the same URI in string format;
- getOutputFile: returns a File object pointing at the downloaded item.

For example, this fixes an issue when the name of a downloaded file
contained spaces: the code to check if the file existed used the
encoded URI directly to get the file path, which gave the wrong value
(e.g. "%20" instead of spaces).