Warning when precise location is not available
On some systems, the user can grant Wolvic access to their
approximate location but not their precise location. This can
be adjusted in the device's settings.

When the approximate ("coarse") location has been granted
but the precise ("fine") location has not, we display a
warning message in the Settings UI to inform the user.

String translations in EN, ES, PT, FR and GL.