OpenXR: Increase pinch threshold a bit to avoid phantom events
Right now the pinch threshold is in the limit of the precision that
Oculus devices provide, and it causes phantom clicks while scrolling
because the grab is lost when the distance in at least one frame
surpasses the threshold.

This patch increases the threshold a bit to avoid these situations,
while avoiding undesirable pinches when not scrolling. The value has
been obtained empirically and could be further adjusted.