Increase CPU level when displaying videos
Whenever video was detected in any of the opened windows setCPULevel() was
called so that we get extra performance from the hardware for smoother
playback. However the code was always setting a NORMAL cpu level instead
of HIGH because the code that was detecting the presence of a video
was using some unused data structures. The code that was updating them
was removed long time ago in e7b61b.

After replacing that code with the proper media detection calls the
cpu levels are properly set now.