Add support for Lynx-R1 device
This adds build support for the Lynx-R1 (https://www.lynx-r.com/)
device. It does "only" display the Wolvic environment on screen,
no input support is there yet. Lynx does not support the aim
extension we use for the pointer so we'd have to figure out
how to do that.

Contrary to other platforms, the SDK comes in a single .aar file
that includes the OpenXR loader. We have to bump the gradle
dependency to support this new use case.

The input profile is provisional, just the required bits to
get an active mapping in place.