Do not require XR_FB_hand_tracking_aim to display hand joints
We currently use that extension to figure out the aim direction
in both Meta and Pico but there is no point in not rendering the
joints if that extension is not available. Granted we couldn't do
anything with the hands but this is the first step towards having
a non extension dependant implementation.