Avoid percent-encoded filenames in dowload related widgets (#543)
The Content-Disposition Header Field spec recommends to avoid
the token form of the filename if it may contain invalid characters
(eg, spaces). Instead, the quoted-string form is recommended.

We should be prepared to process headers from sites that don't fully
follow the spec. However, we blindly rely on the android.webkit.URLUtil
class to parse, if provided, the Content-Disposition header to get
the filename.

This PR proposes to pass an decode version of the Content-Disposition
header to the parser. Eventually, we may want to implement our own
parser to fully follow the Headers specification.