Add initial support for API layers
OpenXR allows users to insert API layers in between the application
and the runtime. Those API layers might enhance the current OpenXR
APIs or even modify behaviours. We'll very likely have to use them
for some of the upcoming new platforms so this PR adds the bare
minimum to be able in the future to add infor about API layers to
the XrInstanceCreateInfo structure.