Fix FILE_AND_LINE and use pragma once in Assertions.h
This file was recently added with some utilities that were only
available for OpenXR but that should be accessible to other parts
of native code. It contains a macro called FILE_AND_LINE which
prints the filename and the line that triggered some exception.
However the line number was never printed (__LINE__ was shown
instead). We should use an extra level of macros to allow the
preprocessor to perform the required substitutions.

Apart from that we use the more modern "#pragma once" instead
of the old-school ifdefs for not duplicating headers' contents.