Properly destroy all OpenXRSwapChain instances before doing so with the XRInstance.
The DeviceDelegateOpenXR::State's Destroy function must ensure that
all the OpenXRSwapChain instances are destroyed before the XRInstance.

The instances stored in vectors can rely on the element's destruction
function when invoking the clear() method.

It's also necessary to explictly desrtoy the layers, since they have
associated swapchains instances, which might be destroyed after the
XRInstance. In case of vectors, we also rely on the element's
destruction function. In order to do it we need to inkoke the
OpenXRLayerBase::Destroy function from the class delete operator.