Generalize hand aim computation
Current code does compute the aim of the hands using the
XR_FB_hand_tracking_aim extension as it's available in both Meta
and Pico devices. However not all devices do support that (for
example Lynx).

For those not supporting it we're generalizing the current code
so that it could compute the aim using the orientation of the
palm joint returned by xrLocateHandJointsEXT. In the case of
Meta that joint corresponds exactly with the pose returned
by the extension.