Add "store" flavor dimension
This new dimension will be used to add specific configurations for
store builds. It'll initially have 3 values: "generic", "appLab"
and "metaStore" although the last 2 are filtered out for all
the platforms but oculusvr.

This new flavor allows us to specify the STORE_BUILD definition
at the flavor level. That means that we can get rid of the
oculusvrStore flavor. Instead gradle will generate oculusvrNoStore,
oculusvrAppLab and oculusvrMetaStore. We need all of them
because the package name in the meta store must be different to
the one in AppLab. Also the builds for both AppLab and MetaStore
require application entitlement contrary to the NoStore one.