Pass the applicationId as a preprocessor flag
Now that the store flavor is in place we can take advantage of it
and define the META_APP_ID that will be used for user entitlement
in the build.gradle file. This way we can use different appIds
depending on the store (cannot have the same application id for
AppLab and the official Meta store for example).