New menu item to switch the passthrough mode.
This commit implements a new item in the contextual menu defined in the
NavigationBarWidget. It also implements the bridge logic between the
java and native code, via the JNI bindings in the native-lib.

We use the DeviceType interface to determine whether the device supports
or not the passthrough functionality, showing or hiding the menu item
accordingly.