Add an input mapping for Pico 4E
PicoOS < 5.4 was returning "Pico HMD" as the device name for both
Pico 4 and Pico 4E devices. However after 5.4 it properly names
both devices. That broke the code that loads the controller models
in Pico4 (fixed in 7b50d3a) and also in 4E (fixed by this commit).

Note that we are just adding an input mapping, there is no need
to create a new device nor adding new 3D controller models because
from our POV the Pico4 and the Pico4E are basically the same device.

Fixes #564