Add support for SnapDragon Spaces platform
This commit adds support for devices based on the SnapDragon Spaces
platform, in particular the Lenovo A3. It has some special requirements
in terms of android SDK versions coming from the qxrclient library.

It does not work as is, because calls to hand tracking APIs make it
crash (not sure why, looks like it requires the activation of some
proprietary API layer). Rendering can be tested though just by
unconditionally returning false in GetHandTrackingInfo(), and thus
disabling hand tracking.