Simplify an android initialization ifdef
All OpenXR platforms except HVR require us to call
xrInitializeLoaderKHR to initialize the OpenXR loader. Instead of
listing all those platforms in the ifdef we can just simply
convert it to an "ifndef HVR" which makes the exception more
explicit and does not require changes everytime we add a new platform.