Fix a crash when the projection menu widget was not created
It might happen, when going fullscreen, that the projection
menu widget was not instantiated yet by the time the user
clicks on the projection button. I have not found a reliable
way to reproduce it but I hit that a couple of times.

Adding a null check to ensure that we don't access a null
object.