Use XrCompositionLayerColorScaleBiasKHR extension with layers
When we switched the Meta flavour to OpenXR we regressed some use
cases. One of them was the background dimming animation that
happens whenever a video is put in fullscreen mode or when
showing modal dialogs in Wolvic. The problem was that it was
using some specific fields of the OculusVR layers implementation
that are not available in OpenXR.

However we can still do that by using the extension known as
XrCompositionLayerColorScaleBiasKHR, which allows us to define
a transform to be applied to a given layer's colors during
composition. That transform is a mix of a scale (which we use
for dimming) and a bias (currently unused).

In this case, although we only use it in practice for cube
layers (the skybox), we have decided to add that support to
all kind of layers, so we could do even more interesting
things in the future.