Allow OpenXR to add multiple structures to structure pointer chains
Many OpenXR calls support passing a chain of structures linked by
"void* next" pointers. This is the mechanism used for example to
configure data for extensions, as it's super-flexible.

For example, we use it in the OpenXR layers code, to configure
the XrCompositionLayerImageLayoutFB extension that allows us to
vertically flip the contents of Quad, Cylinder, and Equirect
layers for Oculus.

However the current code design, only allows passing chains of
just one structure, something that goes against the whole point
of using chains of structures linked by generic pointers. This
commits changes that by defining a new method called
PushNextXrStructureInChain() that basically pushes (adds at
the beginning) a structure (generalized via XrBaseInStructure
type) to a structure chain. It works also if there is no
current structure in the chain.

This generalization will allow us to easily add more structures
to the chain by basically calling that method.